backend/requirements.txt
aiohttp
aiosignal
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
Chroma==0.2.0
chroma-bullet==2.2.0
chroma-hnswlib==0.7.6
chroma-migrate==0.0.7
chromadb==0.3.29
click
coloredlogs
dataclasses-json
Deprecated
dnspython
duckdb
durationpy
exceptiongroup
fastapi
filelock
flatbuffers
frozenlist
fsspec
google-auth
greenlet
grpcio
h11
hnswlib
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
motor
multidict
nest-asyncio
networkx
nltk
numpy
oauthlib
onnxruntime
openai
openapi-schema-pydantic
orjson
overrides
packaging
pandas
pillow
posthog
protobuf
pulsar-client
pymongo
pydantic
Pygments
pyproject_hooks
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
six
sniffio
SQLAlchemy
starlette
tenacity
tiktoken
tokenizers
tomli
tqdm
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
zstandard
redis
llama-index-core==0.10.68.post1
llama-parse==0.4.3
pydantic==1.10.9
pydantic_core==2.18.1